Replaced int with i32 in test_utils.

+++ test_utils/block_utils.m
@@ -1,28 +1,28 @@
-#import <Foundation/Foundation.h>
+#include <stdint.h>
 
-typedef NSInteger (^IntBlock)();
-typedef NSInteger (^AddBlock)(NSInteger);
+typedef int32_t (^IntBlock)();
+typedef int32_t (^AddBlock)(int32_t);
 
 IntBlock get_int_block() {
-    return ^{ return (NSInteger)7; };
+    return ^{ return (int32_t)7; };
 }
 
-IntBlock get_int_block_with(NSInteger i) {
+IntBlock get_int_block_with(int32_t i) {
     return [^{ return i; } copy];
 }
 
 AddBlock get_add_block() {
-    return ^(NSInteger a) { return a + 7; };
+    return ^(int32_t a) { return a + 7; };
 }
 
-AddBlock get_add_block_with(NSInteger i) {
-    return [^(NSInteger a) { return a + i; } copy];
+AddBlock get_add_block_with(int32_t i) {
+    return [^(int32_t a) { return a + i; } copy];
 }
 
-NSInteger invoke_int_block(IntBlock block) {
+int32_t invoke_int_block(IntBlock block) {
     return block();
 }
 
-NSInteger invoke_add_block(AddBlock block, NSInteger a) {
+int32_t invoke_add_block(AddBlock block, int32_t a) {
     return block(a);
 }
